+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Note the first digit of every adapter version corresponds to the major version o
 Adapters are compatible with any Chartboost Mediation SDK version within that major version.
 
 ### 4.3.0.0.0
+- Uses BidMachine's `getExtrasParam()` method when bidding.
 - This version of the adapter has been certified with BidMachine Ads SDK 3.0.0.
 
 ### 4.2.7.0.0